[Issue #9590] Move fundingOptions and categoryOptions to constants/opportunity
